[step2] refactor : web UI를 위한 ConsoleView 분리
